requirements.txt
django
pillow
uuid
emailMessage
dotenv
firebase_admin
smtplib
pyrebase4
folium
